[Workflow Model] add support for model extensions
Add the ability for model extensions, these are sets of hook functions that allow for custom overrides and extensions during model functions. For example, if we wanted to add text search support to the submissions listing functionaility we could hook into modify the listing method and add additional filtering parameters.