[DS Theme] implementation of autocomplete using popper tooltip for the menu overlay
This is a re-implementation (with improvements) of the existing autocomplete implementation based around 'react-autocomplete'. By using the react-popper-tooltip library as the underlying implementation for the popover menu we get better user interactiona and placement functionaility. Also, shifted some of the logic commonly utilised inside the base implementation (debounced value etc).